LayoutTests/fast/forms/state-restore-per-form-expected.txt

CONSOLE MESSAGE: line 7: Form state vector:
CONSOLE MESSAGE: line 7: , text, No owner, 1, visited, 
CONSOLE MESSAGE: line 7: before, text, No owner, 1, form3-outer1-modified, 
CONSOLE MESSAGE: line 7: before, text, No owner, 1, noowner1-modified, 
CONSOLE MESSAGE: line 7: before, text, No owner, 1, form2-outer1-modified, 
CONSOLE MESSAGE: line 7: common-name1, text, state-restore-per-form-back.html #0, 0, common-name1, 
CONSOLE MESSAGE: line 7: text, state-restore-per-form-back.html #0, 1, form1-control2-modified, common-name1, 
CONSOLE MESSAGE: line 7: text, http://example.com/foo.cgi#bar #0, 1, query-changed-control1-modified, username, 
CONSOLE MESSAGE: line 7: text, http://example.com/foo.cgi#bar #1, 1, username-modified, common-name1, 
CONSOLE MESSAGE: line 7: text, http://example.com/foo.cgi#bar #1, 1, form3-control2-modified, after, 
CONSOLE MESSAGE: line 7: text, No owner, 1, form3-outer2-modified, after, 
CONSOLE MESSAGE: line 12: text, No owner, 1, form2-outer2-modified, 
CONSOLE MESSAGE: line 13: 
Confirm we can restore correctly even if the order of forms are changed:
PASS $("form1-control1").value is "initial"
PASS $("form1-control2").value is "form1-control2-modified"
PASS $("form3-control2").value is "form3-control2-modified"

Confirm we ignore the query part of action URLs:
PASS $("form2-control1").value is "query-changed-control1-modified"

Confirm an additional control in the previous form doesn't take a state for the next form:
PASS $("form3-control1").value is "username-modified"
PASS $("form2-added").value is "initial"

Associated controls with form attributes are correctly handled:
PASS $("form3-outer1").value is "form3-outer1-modified"
FAIL $("form2-outer1").value should be form2-outer1-modified. Was noowner1-modified.
FAIL $("noowner2").value should be initial. Was form3-outer2-modified.
FAIL $("form3-outer2").value should be form3-outer2-modified. Was form2-outer2-modified.
FAIL $("form2-outer2").value should be form2-outer2-modified. Was initial.
PASS successfullyParsed is true

TEST COMPLETE
